ChartEx R6 Class for Extended Spreadsheet Charts

An R6 class to create and manipulate Excel Extended Charts (ChartEx), including Waterfall, Sunburst, Treemap, and Region Maps, which are not supported by standard Office Open XML chart types.

Details

This class uses `xml2` to manipulate the underlying XML structure and integrates with `openxlsx2` for workbook generation.

Public fields

xml

The raw XML object for the chart space.

series_data

A list containing all added series definitions.

chart_title

The title text.

chart_title_style

List of styling parameters for the title.

x_title

The X-axis title text.

x_title_style

List of styling parameters for the X-axis title.

y_title

The Y-axis title text.

y_title_style

List of styling parameters for the Y-axis title.

x_axis_style

List of styling and axis parameters for the X-axis.

y_axis_style

List of styling and axis parameters for the Y-axis.

x_numfmt

Number format for X-axis.

y_numfmt

Number format for Y-axis.

legend_params

Parameters for legend positioning and style.

data_label_params

Parameters for data labels.

chart_area_style

Styling for the outer chart area.

plot_area_style

Styling for the inner plot area.

Methods

Public methods

ChartEx$new()

ChartEx$set_chart_style()

ChartEx$set_plot_style()

ChartEx$set_chart_title()

ChartEx$set_x_title()

ChartEx$set_y_title()

ChartEx$set_legend_style()

ChartEx$set_data_label_style()

ChartEx$set_x_axis()

ChartEx$set_y_axis()

ChartEx$add_series()

ChartEx$render()

ChartEx$clone()

Method new()

Create a new ChartEx object.

Usage

ChartEx$new()

Returns

A new `ChartEx` object.

Method set_chart_style()

Set chart area styling.

Usage

ChartEx$set_chart_style(fill = NULL, line = NULL, line_width = 1)

Arguments

fill

Fill color (hex or wbColour).

line

Line/border color.

line_width

Width of the border.

Method set_plot_style()

Set plot area styling.

Usage

ChartEx$set_plot_style(fill = NULL, line = NULL, line_width = 1)

Arguments

fill

Fill color (hex or wbColour).

line

Line/border color.

line_width

Width of the border.

Method set_chart_title()

Set the chart title and style.

Usage

ChartEx$set_chart_title(text, ...)

Arguments

text

Title text.

...

Styling (sz, b, color, font).

Method set_x_title()

Set the X-axis title and style.

Usage

ChartEx$set_x_title(text, ...)

Arguments

text

Title text.

...

Styling (sz, b, color, font).

Method set_y_title()

Set the Y-axis title and style.

Usage

ChartEx$set_y_title(text, ...)

Arguments

text

Title text.

...

Styling (sz, b, color, font).

Method set_legend_style()

Set legend properties.

Usage

ChartEx$set_legend_style(
  pos = "t",
  align = "ctr",
  overlay = FALSE,
  sz = NULL,
  name = NULL,
  bold = NULL,
  italic = NULL,
  color = NULL
)

Arguments

pos

Position (t, b, l, r, none).

align

Alignment (ctr, min, max).

overlay

Logical; overlay legend on chart.

sz

Size of font.

name

Name of font.

bold

Logical.

italic

Logical.

color

Hex color.

Method set_data_label_style()

Set data label properties.

Usage

ChartEx$set_data_label_style(
  show = TRUE,
  pos = "outEnd",
  sz = NULL,
  name = NULL,
  bold = NULL,
  italic = NULL,
  color = NULL,
  numfmt = NULL
)

Arguments

show

Logical.

pos

Position (outEnd, inEnd, ctr, etc).

sz

Font size.

name

Font name.

bold

Logical.

italic

Logical.

color

Hex color.

numfmt

Excel number format string.

Method set_x_axis()

Set X-axis tick label style, number format, and parameters.

Usage

ChartEx$set_x_axis(numfmt = NULL, ...)

Arguments

numfmt

Excel number format.

...

Styling parameters (sz, color, b, font) and axis parameters (gap_width, min, max, grid_color).

Method set_y_axis()

Set Y-axis tick label style, number format, and parameters.

Usage

ChartEx$set_y_axis(numfmt = NULL, ...)

Arguments

numfmt

Excel number format.

...

Styling parameters (sz, color, b, font) and axis parameters (gap_width, min, max, grid_color).

Method add_series()

Add a data series to the chart.

Usage

ChartEx$add_series(
  header = NULL,
  data,
  cat = NULL,
  type = "waterfall",
  fill_color = "auto",
  line_color = NULL,
  line_width = 1,
  subtotals = NULL
)

Arguments

header

Cell range for the series name.

data

Cell range for the numeric values.

cat

Cell range for the category labels.

type

Type of chart (waterfall, sunburst, treemap, regionMap).

fill_color

Hex color or "auto".

line_color

Border color.

line_width

Border width.

subtotals

Numeric vector of indices to treat as subtotals (Waterfall only).

Method render()

Render the internal XML for writing to a file.

Usage

ChartEx$render(id_start = 1)

Arguments

id_start

Numeric starting ID for XML data references.

Returns

A list containing the XML and attribute mappings.

Method clone()

The objects of this class are cloneable with this method.

Usage

ChartEx$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.